changed loading wav method

+++ mel2samp.py
@@ -40,6 +40,7 @@ from tacotron2.layers import TacotronSTFT
 
 MAX_WAV_VALUE = 32768.0
 
+
 def files_to_list(filename):
     """
     Takes a text file of filenames and makes a list of filenames
@@ -50,10 +51,13 @@ def files_to_list(filename):
     files = [f.rstrip() for f in files]
     return files
 
+
 def load_wav_to_torch(full_path, sampling_rate):
     data = librosa.core.load(full_path, sr=sampling_rate)[0]
     data = data / np.abs(data).max() * 0.999
     return torch.FloatTensor(data.astype(np.float32))
+
+
 # def load_wav_to_torch(full_path):
 #     """
 #     Loads wavdata into torch array
@@ -64,7 +68,8 @@ def load_wav_to_torch(full_path, sampling_rate):
 int16_max = (2 ** 15) - 1
 audio_norm_target_dBFS = -30
 
-def preprocess_wav(fpath_or_wav, source_sr = None, sampling_rate=16000):
+
+def preprocess_wav(fpath_or_wav, source_sr=None, sampling_rate=16000):
     # Load the wav from disk if needed
     if isinstance(fpath_or_wav, str):
         wav, source_sr = librosa.load(fpath_or_wav, sr=None)
@@ -80,6 +85,7 @@ def preprocess_wav(fpath_or_wav, source_sr = None, sampling_rate=16000):
     wav = wav / np.abs(wav).max() * 0.999
     # wav = trim_long_silences(wav)
     return wav
+
 
 def normalize_volume(wav, target_dBFS, increase_only=False, decrease_only=False):
     if increase_only and decrease_only:
@@ -97,6 +103,7 @@ class Mel2Samp(torch.utils.data.Dataset):
     This is the main class that calculates the spectrogram and returns the
     spectrogram, audio pair.
     """
+
     def __init__(self, training_files, segment_length, filter_length,
                  hop_length, win_length, sampling_rate, mel_fmin, mel_fmax, num_workers, npy_dir):
         self.audio_files = files_to_list(training_files)
@@ -133,7 +140,7 @@ class Mel2Samp(torch.utils.data.Dataset):
         if audio.size(0) >= self.segment_length:
             max_audio_start = audio.size(0) - self.segment_length
             audio_start = random.randint(0, max_audio_start)
-            audio = audio[audio_start:audio_start+self.segment_length]
+            audio = audio[audio_start:audio_start + self.segment_length]
         else:
             audio = torch.nn.functional.pad(audio, (0, self.segment_length - audio.size(0)), 'constant').data
         mel = self.get_mel(audio)
@@ -144,6 +151,7 @@ class Mel2Samp(torch.utils.data.Dataset):
 
     def __len__(self):
         return len(self.audio_files)
+
 
 # ===================================================================
 # Takes directory of clean audio and makes directory of spectrograms
@@ -190,16 +198,18 @@ if __name__ == "__main__":
     #     # list(tqdm(pool.imap(preprocess_speaker, speaker_dirs), dataset_name, len(speaker_dirs),
     #     list(pool.map(local_mel2samp, filepaths))
 
-
-
     for filepath in filepaths:
         filepath = filepath.split("|")[0]
+        filename = os.path.basename(filepath)
+        new_filepath = args.output_dir + '/' + filename + '.pt'
+        if os.path.isfile(new_filepath):
+            print("skip", new_filepath)
+            continue
         audio = preprocess_wav(filepath, sampling_rate=args.sampling_rate)
         # audio = load_wav_to_torch(filepath, args.sampling_rate)
         audio = torch.FloatTensor(audio.astype(np.float32))
 
         melspectrogram = mel2samp.get_mel(audio)
-        filename = os.path.basename(filepath)
-        new_filepath = args.output_dir + '/' + filename + '.pt'
+
         print(new_filepath)
         torch.save(melspectrogram, new_filepath)
